+++ storybook/src/stories/person/avatar-example.stories.tsx
@@ -3,10 +3,10 @@ import {
   PersonProvider,
   PersonAvatar,
   AvatarSize,
-  PersonAvailability,
   PersonAccountType,
 } from '@equinor/fusion-react-person/src';
 import type { PersonAvatarProps } from '@equinor/fusion-react-person/src';
+import { createResolve } from './PersonResolve';
 
 export default {
   title: 'Examples/Person/Person Avatar',
@@ -60,14 +60,6 @@ export default {
         defaultValue: { summary: false },
       },
     },
-    availability: {
-      description: 'Availability of the person',
-      control: 'select',
-      options: PersonAvailability,
-      table: {
-        type: { summary: 'PersonAvailability' },
-      },
-    },
     accountType: {
       description: 'Employment type of the person',
       control: 'select',
@@ -80,47 +72,11 @@ export default {
 } as Meta;
 
 export type AvatarProps = PersonAvatarProps & {
-  availability?: PersonAvailability;
   accountType: PersonAccountType;
 };
 
-const createResolve = (accountType: PersonAccountType, availability?: PersonAvailability) => ({
-  getImageByAzureId: async (azureId: string) => {
-    await new Promise((resovle) => setTimeout(resovle, 3000));
-    return await Promise.resolve({
-      azureId: azureId,
-      name: 'Albert Einstein',
-      pictureSrc: 'https://i.imgur.com/GcZeeXX.jpeg',
-      accountType: PersonAccountType.Employee,
-    });
-  },
-  getImageByUpn: async (_upn: string) => {
-    await new Promise((resovle) => setTimeout(resovle, 3000));
-    return await Promise.resolve({
-      name: 'Albert Einstein',
-      pictureSrc: 'https://i.imgur.com/GcZeeXX.jpeg',
-      accountType,
-    });
-  },
-  getPresence: async (azureId: string) => {
-    await new Promise((resolve) => setTimeout(resolve, 2000));
-    return {
-      azureId,
-      availability: availability ?? PersonAvailability.Offline,
-    };
-  },
-  getDetails: async (azureId: string) => {
-    await new Promise((resolve) => setTimeout(resolve, 1000));
-    return {
-      azureId,
-      accountType,
-      pictureSrc: 'https://prod-images.tcm.com/Master-Profile-Images/BurtReynolds.jpg',
-    };
-  },
-});
-
-export const Component: Story<AvatarProps> = ({ accountType, availability, ...props }: AvatarProps) => (
-  <PersonProvider resolve={createResolve(accountType, availability)}>
+export const Component: Story<AvatarProps> = ({ accountType, ...props }: AvatarProps) => (
+  <PersonProvider resolve={createResolve(accountType)}>
     <PersonAvatar {...props} />
   </PersonProvider>
 );
@@ -129,12 +85,11 @@ Component.args = {
   size: AvatarSize.Medium,
   clickable: false,
   disabled: false,
-  availability: PersonAvailability.Available,
   accountType: PersonAccountType.Employee,
 };
 
-export const DataSource: Story<AvatarProps> = ({ availability, ...props }: AvatarProps) => (
-  <PersonProvider resolve={createResolve(PersonAccountType.External, availability)}>
+export const DataSource: Story<AvatarProps> = (props: AvatarProps) => (
+  <PersonProvider resolve={createResolve(PersonAccountType.External)}>
     <PersonAvatar
       {...props}
       dataSource={{
@@ -146,7 +101,7 @@ export const DataSource: Story<AvatarProps> = ({ availability, ...props }: Avata
 );
 
 export const Size: Story<{ sizes: Array<AvatarProps['size']> }> = (props: { sizes: Array<AvatarProps['size']> }) => (
-  <PersonProvider resolve={createResolve(PersonAccountType.Employee, PersonAvailability.DoNotDisturb)}>
+  <PersonProvider resolve={createResolve(PersonAccountType.Employee)}>
     <div style={{ display: 'flex', alignItems: 'center', gap: 20 }}>
       {props.sizes.map((size) => (
         <PersonAvatar key={size} size={size} upn="mail@mail.com" />
@@ -156,19 +111,18 @@ export const Size: Story<{ sizes: Array<AvatarProps['size']> }> = (props: { size
 );
 Size.args = { sizes: [AvatarSize.Large, AvatarSize.Medium, AvatarSize.Small, AvatarSize.XSmall] };
 
-export const Clickable: Story<AvatarProps> = ({ availability, ...props }: AvatarProps) => (
-  <PersonProvider resolve={createResolve(PersonAccountType.Employee, availability)}>
+export const Clickable: Story<AvatarProps> = (props: AvatarProps) => (
+  <PersonProvider resolve={createResolve(PersonAccountType.Employee,)}>
     <PersonAvatar {...props} />
   </PersonProvider>
 );
 Clickable.args = {
   azureId: '1234',
   clickable: true,
-  availability: PersonAvailability.Away,
 };
 
-export const CardOnHover: Story<AvatarProps> = ({ availability, ...props }: AvatarProps) => (
-  <PersonProvider resolve={createResolve(PersonAccountType.External, availability)}>
+export const CardOnHover: Story<AvatarProps> = (props: AvatarProps) => (
+  <PersonProvider resolve={createResolve("External")}>
     <PersonAvatar {...props} />
   </PersonProvider>
 );
@@ -178,8 +132,8 @@ CardOnHover.args = {
   showFloatingOn: "hover"
 };
 
-export const CardOnClick: Story<AvatarProps> = ({ availability, ...props }: AvatarProps) => (
-  <PersonProvider resolve={createResolve(PersonAccountType.External, availability)}>
+export const CardOnClick: Story<AvatarProps> = (props: AvatarProps) => (
+  <PersonProvider resolve={createResolve(PersonAccountType.External)}>
     <PersonAvatar {...props} />
   </PersonProvider>
 );
@@ -189,8 +143,8 @@ CardOnClick.args = {
   showFloatingOn: "click"
 };
 
-export const Disabled: Story<AvatarProps> = ({ availability, ...props }: AvatarProps) => (
-  <PersonProvider resolve={createResolve(PersonAccountType.External, availability)}>
+export const Disabled: Story<AvatarProps> = (props: AvatarProps) => (
+  <PersonProvider resolve={createResolve(PersonAccountType.External)}>
     <div>
       <PersonAvatar {...props} />
     </div>
@@ -199,5 +153,4 @@ export const Disabled: Story<AvatarProps> = ({ availability, ...props }: AvatarP
 Disabled.args = {
   azureId: '1234',
   disabled: true,
-  availability: PersonAvailability.DoNotDisturb,
 };